move codenode.py and keyed_region.py to global scope
