+++ angr/analyses/binary_optimizer.py
@@ -7,7 +7,7 @@ from . import Analysis, register_analysis
 
 from ..knowledge_base import KnowledgeBase
 from .. import SIM_PROCEDURES
-from ..knowledge import HookNode
+from ..codenode import HookNode
 from ..sim_variable import SimConstantVariable, SimRegisterVariable, SimMemoryVariable, SimStackVariable
 
 l = logging.getLogger("angr.analyses.binary_optimizer")
