--- angr/knowledge/__init__.py
@@ -1,2 +1,0 @@
-from .codenode import CodeNode, BlockNode, HookNode
-
